feat: merge vault upgrade and core reaudit into one bnb proposal
