+++ simulations/vip-665/abi/ResilientOracle.json
